Conflicts:
	ArchipelAgent/archipel-agent/install/etc/archipel/archipel.conf
	ArchipelClient/Resources/archipel-commit.js

+++ ArchipelAgent/archipel-agent-virtualmachine-oomkiller/archipelagentvirtualmachineoomkiller/oomkiller.py
@@ -21,10 +21,11 @@
 
 import commands
 import sqlite3
-import xmpp
+import os
 
 from archipelcore.archipelPlugin import TNArchipelPlugin
 from archipelcore.utils import build_error_iq
+from archipelcore import xmpp
 
 
 ARCHIPEL_NS_OOM_KILLER = "archipel:vm:oom"
@@ -159,7 +160,10 @@ class TNOOMKiller (TNArchipelPlugin):
         """
         try:
             pid = int(commands.getoutput("ps -ef | grep kvm | grep %s | grep -v grep" % self.entity.uuid).split()[1])
-            f = open("/proc/%d/oom_adj" % pid, "w")
+            if os.path.isfile("/proc/%d/oom_score_adj" % pid):
+                f = open("/proc/%d/oom_score_adj" % pid, "w")
+            else:
+                f = open("/proc/%d/oom_adj" % pid, "w")
             f.write(str(adjust))
             f.close()
         except Exception as ex:
